cyber/phisingTool/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
email-validator==2.0.0
Jinja2==3.1.2
Flask-Mail==0.9.1
SQLAlchemy==2.0.21
Werkzeug==2.3.7
python-dotenv==1.0.0